% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_install.R
\name{load_install}
\alias{load_install}
\title{a convenience function to install a package if it isn't already}
\usage{
load_install(pkg)
}
\description{
a convenience function to install a package if it isn't already
}

