% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UhUm.R
\name{UhUm}
\alias{UhUm}
\alias{UhUm-package}
\alias{UmUh}
\title{A package with Um and Uh Data}
\description{
A package with Um and Uh Data
}
\examples{
browseVignettes("UhUm")
vignette("contributing")
vignette("PNCdata")
}

