% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{um_PNC}
\alias{um_PNC}
\title{Uh and Um data from the PNC}
\format{a data frame with 19 columns. Speech chunks were defined by
continuous speech bordered on each side by a pause of 200 miliseconds or more
 \describe{
   \item{idstring}{Unique id string for each speaker}
   \item{word}{\code{UM} or \code{UH}}
   \item{start_time}{Onset time of filled pause, in seconds}
   \item{end_time}{Offset time of filled pause, in seconds}
   \item{vowel_start}{Onset time of filled pause vowel}
   \item{vowel_end}{Offset time of filled pause vowel}
   \item{nasal_start}{Onset time of nasal, if \code{UM}}
   \item{nasal_end}{Offset time of nasal, if \code{UM}}
   \item{chunk_start}{Onset time of speech chunk}
   \item{chunk_end}{Offset time of speech chunk}
   \item{nwords}{Total number of words from this speaker}
   \item{sex}{Sex of speaker}
   \item{year}{Year of interview}
   \item{age}{Age of speaker at time of interview}
   \item{ethnicity}{A number of single character codes for speakers' ethnicity}
   \item{schooling}{Number of years of schooling}
   \item{transcribed}{Total number of transcribed seconds of interview}
   \item{total}{Total length of recording}
   \item{nvowels}{Number of measured stressed vowels (irrelevant to UM/UH)}
 }}
\source{
Philadelphia Neighborhood Corpus
}
\usage{
um_PNC
}
\description{
Uh and Um data from the PNC
}
\keyword{datasets}

